Synced defs.hpp with godot's typedefs.h
