+++ include/godot_cpp/core/binder_common.hpp
@@ -32,6 +32,7 @@
 
 #include <gdextension_interface.h>
 
+#include <godot_cpp/core/defs.hpp>
 #include <godot_cpp/core/method_ptrcall.hpp>
 #include <godot_cpp/core/type_info.hpp>
